Clean up owncloud client factories
