--- owncloudComLibrary/src/main/java/com/owncloud/android/lib/common/OwnCloudClientManager.java
@@ -1,52 +1,0 @@
-/* ownCloud Android Library is available under MIT license
- *   Copyright (C) 2019 ownCloud GmbH.
- *
- *   Permission is hereby granted, free of charge, to any person obtaining a copy
- *   of this software and associated documentation files (the "Software"), to deal
- *   in the Software without restriction, including without limitation the rights
- *   to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- *   copies of the Software, and to permit persons to whom the Software is
- *   furnished to do so, subject to the following conditions:
- *
- *   The above copyright notice and this permission notice shall be included in
- *   all copies or substantial portions of the Software.
- *
- *   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- *   EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- *   MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- *   NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
- *   BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
- *   ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- *   CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- *   THE SOFTWARE.
- *
- */
-
-package com.owncloud.android.lib.common;
-
-import android.accounts.AuthenticatorException;
-import android.accounts.OperationCanceledException;
-import android.content.Context;
-
-import com.owncloud.android.lib.common.accounts.AccountUtils.AccountNotFoundException;
-
-import java.io.IOException;
-
-/**
- * Manager to create and reuse OwnCloudClient instances to access remote OC servers.
- *
- * @author David A. Velasco
- * @author masensio
- * @author Christian Schabesberger
- */
-
-public interface OwnCloudClientManager {
-
-    OwnCloudClient getClientFor(OwnCloudAccount account, Context context) throws AccountNotFoundException,
-            OperationCanceledException, AuthenticatorException, IOException;
-
-    OwnCloudClient removeClientFor(OwnCloudAccount account);
-
-    void saveAllClients(Context context, String accountType) throws AccountNotFoundException, AuthenticatorException, IOException,
-            OperationCanceledException;
-}
