+++ owncloudComLibrary/src/main/java/com/owncloud/android/lib/common/SingleSessionManager.java
@@ -1,5 +1,5 @@
 /* ownCloud Android Library is available under MIT license
- *   Copyright (C) 2019 ownCloud GmbH.
+ *   Copyright (C) 2020 ownCloud GmbH.
  *
  *   Permission is hereby granted, free of charge, to any person obtaining a copy
  *   of this software and associated documentation files (the "Software"), to deal
@@ -41,23 +41,35 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 /**
- * Implementation of {@link OwnCloudClientManager}
- * <p>
- * TODO check multithreading safety
- *
  * @author David A. Velasco
  * @author masensio
  * @author Christian Schabesberger
  * @author David González Verdugo
  */
 
-public class SingleSessionManager implements OwnCloudClientManager {
+public class SingleSessionManager {
+
+    private static SingleSessionManager sDefaultSingleton;
+    private static String sUserAgent;
 
     private ConcurrentMap<String, OwnCloudClient> mClientsWithKnownUsername = new ConcurrentHashMap<>();
-
     private ConcurrentMap<String, OwnCloudClient> mClientsWithUnknownUsername = new ConcurrentHashMap<>();
 
-    @Override
+    public static SingleSessionManager getDefaultSingleton() {
+        if (sDefaultSingleton == null) {
+            sDefaultSingleton = new SingleSessionManager();
+        }
+        return sDefaultSingleton;
+    }
+
+    public static String getUserAgent() {
+        return sUserAgent;
+    }
+
+    public static void setUserAgent(String userAgent) {
+        sUserAgent = userAgent;
+    }
+
     public OwnCloudClient getClientFor(OwnCloudAccount account, Context context) throws OperationCanceledException,
             AuthenticatorException, IOException {
 
@@ -100,7 +112,6 @@ public class SingleSessionManager implements OwnCloudClientManager {
                     true);    // TODO remove dependency on OwnCloudClientFactory
             client.setAccount(account);
             HttpClient.setContext(context);
-            client.setOwnCloudClientManager(this);
 
             account.loadCredentials(context);
             client.setCredentials(account.getCredentials());
@@ -126,12 +137,11 @@ public class SingleSessionManager implements OwnCloudClientManager {
         return client;
     }
 
-    @Override
-    public OwnCloudClient removeClientFor(OwnCloudAccount account) {
+    public void removeClientFor(OwnCloudAccount account) {
         Timber.d("removeClientFor starting ");
 
         if (account == null) {
-            return null;
+            return;
         }
 
         OwnCloudClient client;
@@ -140,7 +150,7 @@ public class SingleSessionManager implements OwnCloudClientManager {
             client = mClientsWithKnownUsername.remove(accountName);
             if (client != null) {
                 Timber.v("Removed client for account %s", accountName);
-                return client;
+                return;
             } else {
                 Timber.v("No client tracked for  account %s", accountName);
             }
@@ -149,10 +159,8 @@ public class SingleSessionManager implements OwnCloudClientManager {
         mClientsWithUnknownUsername.clear();
 
         Timber.d("removeClientFor finishing ");
-        return null;
     }
 
-    @Override
     public void saveAllClients(Context context, String accountType) {
         Timber.d("Saving sessions... ");
